requirements.txt
flask==2.2.2
Werkzeug==2.2.2
requests-oauthlib==1.3.0
requests==2.26.0
pyjwt==2.6.0
cryptography==39.0.0